Issue #39 move init code into controller factory
